test: Move existing tests of sort by description to DescriptionField.test.ts
