+++ tests/Query/Filter/DescriptionField.test.ts
@@ -9,6 +9,7 @@ import { fromLine } from '../../TestHelpers';
 import type { FilterOrErrorMessage } from '../../../src/Query/Filter/Filter';
 import { BooleanField } from '../../../src/Query/Filter/BooleanField';
 import { toMatchTaskFromLine } from '../../CustomMatchers/CustomMatchersForFilters';
+import { Sort } from '../../../src/Query/Sort';
 
 window.moment = moment;
 
@@ -242,5 +243,35 @@ describe('search description for Alternation (OR)', () => {
         expect(filter).toMatchTaskFromLine('- [ ] Do stuff waiting');
         expect(filter).toMatchTaskFromLine('- [ ] Do stuff waits');
         expect(filter).toMatchTaskFromLine('- [ ] Do stuff wartet');
+    });
+});
+
+describe('sorting by description', () => {
+    it('sorts correctly by the link name and not the markdown', () => {
+        const one = fromLine({
+            line: '- [ ] *ZZZ An early task that starts with an A; actually not italic since only one asterisk',
+        });
+        const two = fromLine({
+            line: '- [ ] [[Better be second]] with bla bla behind it',
+        });
+        const three = fromLine({
+            line: '- [ ] [[Another|Third it should be]] and not [last|ZZZ]',
+        });
+        const four = fromLine({
+            line: '- [ ] *Very italic text*',
+        });
+        const five = fromLine({
+            line: '- [ ] [@Zebra|Zebra] should be last for Zebra',
+        });
+
+        const expectedOrder = [one, two, three, four, five];
+        expect(
+            Sort.by(
+                {
+                    sorting: [Sort.makeLegacySorting(false, 1, 'description')],
+                },
+                [two, one, five, four, three],
+            ),
+        ).toEqual(expectedOrder);
     });
 });
